+++ init.php
@@ -13,9 +13,10 @@
 
 use Eventum\Event\SystemEvents;
 
+require_once __DIR__ . '/autoload.php';
 require_once __DIR__ . '/globals.php';
 
-if (!file_exists(APP_CONFIG_PATH . '/setup.php') || !filesize(APP_CONFIG_PATH . '/setup.php')) {
+if (Setup::needsSetup()) {
     // redirect to setup
     if (PHP_SAPI === 'cli') {
         throw new RuntimeException('Eventum is not configured');
@@ -30,13 +31,15 @@ ini_set('memory_limit', '512M');
 // prevent session from messing up the browser cache
 ini_set('session.cache_limiter', 'nocache');
 
-require_once APP_CONFIG_PATH . '/config.php';
-require_once APP_PATH . '/autoload.php';
+/**
+ * @deprecated this file won't be loaded in 3.9.0
+ */
+require_once Setup::getConfigPath() . '/config.php';
 
 Misc::stripInput($_POST);
 
 // set default timezone
-date_default_timezone_set(Date_Helper::getDefaultTimezone());
+date_default_timezone_set(Setup::getDefaultTimezone());
 
 Eventum\Monolog\Logger::initialize();
 Language::setup();