+++ lib/eventum/class.setup.php
@@ -13,6 +13,7 @@
 
 use Eventum\Config\Config;
 use Eventum\Config\ConfigPersistence;
+use Eventum\Config\Paths;
 use Eventum\Monolog\Logger;
 use Symfony\Component\Filesystem\Exception\IOException;
 
@@ -44,7 +45,7 @@ class Setup
      */
     public static function getAuthCookie(): array
     {
-        $config = Setup::get();
+        $config = self::get();
 
         return [
             'name' => $config['cookie'],
@@ -58,7 +59,7 @@ class Setup
      */
     public static function getProjectCookie(): array
     {
-        $config = Setup::get();
+        $config = self::get();
 
         return [
             'name' => $config['project_cookie'],
@@ -71,7 +72,7 @@ class Setup
      */
     public static function getBaseUrl(): string
     {
-        return Setup::get()['base_url'];
+        return self::get()['base_url'];
     }
 
     /**
@@ -79,7 +80,7 @@ class Setup
      */
     public static function getRelativeUrl(): string
     {
-        return Setup::get()['relative_url'];
+        return self::get()['relative_url'];
     }
 
     /**
@@ -87,7 +88,7 @@ class Setup
      */
     public static function getHostname(): string
     {
-        return Setup::get()['hostname'];
+        return self::get()['hostname'];
     }
 
     /**
@@ -95,7 +96,7 @@ class Setup
      */
     public static function getAppName(): string
     {
-        return Setup::get()['name'];
+        return self::get()['name'];
     }
 
     /**
@@ -103,7 +104,7 @@ class Setup
      */
     public static function getShortName(): string
     {
-        return Setup::get()['short_name'];
+        return self::get()['short_name'];
     }
 
     /**
@@ -114,7 +115,7 @@ class Setup
      */
     public static function getToolCaption(): string
     {
-        return Setup::get()['tool_caption'];
+        return self::get()['tool_caption'];
     }
 
     /**
@@ -122,7 +123,7 @@ class Setup
      */
     public static function getAnonymousUser(): ?string
     {
-        return Setup::get()['anonymous_user'];
+        return self::get()['anonymous_user'];
     }
 
     /**
@@ -133,7 +134,7 @@ class Setup
      */
     public static function getDefaultTimezone(): string
     {
-        return Setup::get()['default_timezone'] ?? 'UTC';
+        return self::get()['default_timezone'] ?? 'UTC';
     }
 
     /**
@@ -144,7 +145,21 @@ class Setup
      */
     public static function getDefaultWeekday(): int
     {
-        return Setup::get()['default_weekday'];
+        return self::get()['default_weekday'];
+    }
+
+    /**
+     * @return array
+     * @since 3.8.0
+     */
+    public static function getTemplatePaths(): array
+    {
+        $localPath = self::get()['local_path'];
+
+        return [
+            $localPath . '/templates',
+            Paths::APP_TPL_PATH,
+        ];
     }
 
     /**
@@ -228,6 +243,17 @@ class Setup
     public static function getSetupFile(): string
     {
         return self::getConfigPath() . '/setup.php';
+    }
+
+    /**
+     * @since 3.8.0
+     * @return bool
+     */
+    public static function needsSetup(): bool
+    {
+        $setupFile = self::getSetupFile();
+
+        return !file_exists($setupFile) || !filesize($setupFile);
     }
 
     /**