+++ lib/eventum/gettext.php
@@ -12,9 +12,11 @@
  */
 
 // if there is no gettext support built into PHP, or we are running in language compatibility mode include PHP-gettext
+use Eventum\Config\Paths;
+
 if (!function_exists('gettext') || Setup::get()['gettext_mode'] === 'php') {
     /** @noinspection PhpIncludeInspection */
-    require_once APP_PHP_GETTEXT_PATH . '/gettext.inc';
+    require_once Paths::APP_PHP_GETTEXT_PATH . '/gettext.inc';
 
     function ev_gettext($string)
     {
@@ -81,8 +83,6 @@ if (!function_exists('gettext') || Setup::get()['gettext_mode'] === 'php') {
     }
 }
 
-if (defined('APP_PATH')) {
-    _bindtextdomain('eventum', APP_PATH . '/localization/');
-    _bind_textdomain_codeset('eventum', 'UTF-8');
-    _textdomain('eventum');
-}
+_bindtextdomain('eventum', Paths::APP_PATH . '/localization/');
+_bind_textdomain_codeset('eventum', 'UTF-8');
+_textdomain('eventum');