+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Eventum 3.8.x requires MySQL server 5.5.3, see #433
 - Move constants from `config.php` to `setup.php` (@glensc, #629)
 - Replace Misc::getFileList with Symfony Finder (@glensc, #643)
 - Update for new EventDispatcherInterface::dispatch() signature (@glensc, #640)
+- Move global defines to `Paths` class (@glensc, #642)
 
 [3.8.0]: https://github.com/eventum/eventum/compare/v3.7.4...master